+++ libtrackerboy/synth/NoiseGen.cpp
@@ -65,7 +65,7 @@ void NoiseGen::step(uint32_t cycles) {
     // ~((~mLfsr) & 0x1) + 1 -> (mLfsr | (~1)) + 1
     // so we get 0xFF when bit 0 is 0, and 0x00 when bit 0 is 1
 
-    mOutput = (mLfsr | static_cast<uint8_t>(~1)) + 1;
+    mOutput = static_cast<uint8_t>(mLfsr | static_cast<uint8_t>(~1)) + 1;
 }
 
 uint8_t NoiseGen::readRegister() {
